Update authenticator and spawner names and remove unused options
This is a more agressive cleaning of the authenticator and spawners. It
removes all of the systemuser stuff that I could find. So far I've
tested the following configurations (by redeploying and logging in only)

  shib & dockerspawner (I tested the mock accounts)
  dummy & dockerspawner
  pam & dockerspawner (after creating a local account)

I've started trimming out the systemuser stuff from the swiftspawner,
but I'm not really familiar with it so it might still need some work.

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>