Restore docker resource limits and remove dead containers
I think when we shifted to non-root users I over simplified the
jupyterhub global config options. This commit restores the resource
limits (2GB + 1/1 share of a CPU) and tells docker to remove jupyter
containers that have exited. We need to remove containers so that image
updates will propagate to users (eventually).

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>