Update docker image build instructions
We need to use DARGS to override the BASE_CONTAINER in the image build chain to override the default `jupyter/*` images. There are a few ways we can improve this, but the idea is that eventually we want to use the stock `jupyter/scipy-notebook` image as the base for our images.