Add Support for Announcements
We already had support for displaying alerts by setting the
`jupyterhub_announcement` variable. This commit provides the ability to
set general notifications.

It does this by adding a new variable called `jupyterhub_alerts` which
will take the role of `jupyterhub_announcements` and be used for
emergency alerts.

`jupyterhub_announcements` will now be used for info-level messages.