Use same value for ENV across all makefile targets
I was getting confused by some tasks prefixing environment names with
hub- and others not doing so. This change assumes that the content of
ENV is the same across all tasks and corresponds to the names used for
subdirectories of ansible/group_vars and terraform, e.g. ENV=hub-prod or
ENV=hub-iana

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>