Big refactor (#64)
* Start of Large Refactor

This commit is the start of a large refactor. It reorganizes Terraform
templates and modules as well as decouples a lot of Ansible roles. The
end result is still the ability to create an All-in-One JupyterHub
environment, but pieces are in place for other architectures.

* Support for all-in-one and separate ssp

* Continuation of Large Refactor

This commit adds the ability to create two new development environments:

* JupyterHub with separate SimpleSAMLphp
* JupyterHub with built-in SSP and external Stats server

This commit also does a heavy refactor of the previous statistics
configuration.