Moving to Tutor for edX (#67)
This commit changes how edX deployments are created and managed. Instead
of using the edX Ansible playbooks and roles, we are now using Tutor.
Tutor is a Docker-based edX deployment management tool which handles
building Docker images, configuring run-time settings, and managing the
containers as they're running.

Ansible is used to manage the installation of Tutor.

The README and PROCESSES docs have been updated to reflect this.