Initial support for the Sharder
This commit adds initial support for the sharder. There might
be additional changes required, but this commit has been validated
in a running proof of concept.

The sharder will run as its own service "above" the hubs. It also
requires the SimpleSAMLphp service to be decoupled from the hub
and run as its own virtual machine. Once this is in place, the
sharder will appear the same as a hub (with the same landing page)
and will redirect a user to the main authentication page they
see now. Upon successful authentication, the user will return to
the sharder which will then redirect the user to the shard/hub
they are assigned to. Since the user has already authenticated,
they are logged in immediately.