Typo in the defaults for jupyterhub and loosen systemd constraints
There was a typo in the jupyterhub config file which was silently
ignored but meant that single user servers were tidided up during a hub
stoppage.

The storage creation script needs more permissions from systemd to be
able to do it's job, without this correction file ownership is messed
up.

Signed-off-by: Ian Allison <iana@pims.math.ca>