Add a backup task for sqlite and include rrsync config for sharder
These changes add a daily cron job to the sharder to make a backup of
the sharding database (an sqlite file). The script does this locally and
retains one years worth of files (they are tiny). In addition we add the
rrsync role to the sharder to create a remote copy of these files to a
third party service.

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>