Add a jupyterhub service token for prometheus metric collection
The hub now requires authentication for the /hub/metrics endpoint. This
commit adds a `service_token` to the hub for this purpose. It also
restructures `jupyterhub_base_config.py.j2` a little bit to make it
easier to add other services, and we install `pycurl` to improve
performance for hubs with lots of users. Stats can be accessed locally
on each hub with

```
curl -H 'Authorization: token <TOKEN FROM local_vars.yml>' \
  http://127.0.0.1:8000/jupyter/hub/metrics
```

As of this commit, the changes haven't been added to prometheus or
caddy. The caddy changes should be minimal (just add a new endpoint) but
the prometheus changes will update the config so that the token is sent
during scraping.

Signed-off-by: Ian Allison <iana@pims.math.ca>