Redo stats handling for jupyterhub endpoint
We simplify the handling of the jupyterhub/jupyter/hub/metrics
endpoint. We already require authentication for caddy so lets skip it
for the hub. Also add the required scrape config to the stats server.

Some of this work is untested (specificaly deploying a stats
    server) because I think we have a security group
missing, I can only connect to my development dev stats server from
clavius (i.e. not from public IPs).

Signed-off-by: Ian Allison <iana@pims.math.ca>