Be more explicit about USER and USERHASH
Generally we're using USER to refer to the human-readable identifier
(e.g. person@email.com). USERHASH refers to the value as reported by
SimpleSamlPHP. The utility script /usr/local/bin/findhash on the ssp
component should be able to do that transformation and the
`user/findhash` makefile task should report that along with the hub the
user is resident on.

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>