Avoid the zfs module a bug makes it painfully slow
The zfs ansible module should be the best way to interact with our user
storage, but a bug makes it very slow when there are a large number of
users. I think this has been identified as fixed upstream, but since we
want to update ansible (and terraform) in general this bandaid works for
now.

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>