Letsencrypt intermediate CA cert rollover
We were starting to see problems with an expired intermediate CA cert on
some of our services. The underlying cause is given in [this
post](https://letsencrypt.org/docs/dst-root-ca-x3-expiration-september-2021/)
but it seems prudent to rebuild our package images to include the new
ca-certificates packages. I'll open this as a PR and close it as part of
CAN-216.

Signed-off-by: Callysto Sysadmin <sysadmin@callysto.ca>