Update hash calculation
This commit ensures that the sha1 hash user identifier calcualted by the
finduser.php script is the same as the one calculated by simplesaml on
the values provided by the mock authenticator. If the values provided by
the upstream authenticators are unchanged, this should mean that the
user directories would be reconnected correctly. That said, it is not
yet know if the values we get from the upstream authenticators are the
same yet.

Signed-off-by: Cybera Sysadmin <sysadmin@cybera.ca>